+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- cli: Show all todos in `todo list` command
+
 ## [0.9.1] - 2025-09-29
 
 ### Changed