+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.2] - 2025-10-08
+
 ### Changed
 
 - cli: Show all todos in `todo list` command
@@ -321,7 +323,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add events and todos command
 - Add dashboard command
 
-[Unreleased]: https://github.com/yzx9/aim/compare/v0.9.1...HEAD
+[Unreleased]: https://github.com/yzx9/aim/compare/v0.9.2...HEAD
+[v0.9.2]: https://github.com/yzx9/aim/compare/v0.9.1...v0.9.2
 [v0.9.1]: https://github.com/yzx9/aim/compare/v0.9.0...v0.9.1
 [v0.9.0]: https://github.com/yzx9/aim/compare/v0.8.2...v0.9.0
 [v0.8.2]: https://github.com/yzx9/aim/compare/v0.8.1...v0.8.2